+++ src/main/kotlin/com/denisolek/cinema/domain/movie/model/Rating.kt
@@ -8,14 +8,14 @@ import mu.KotlinLogging.logger
 
 data class Rating private constructor(
     val rating: Double,
-    val votes: Long
+    val votes: Int
 ) {
     companion object {
         private val log = logger {}
 
         fun rating(rawRating: String, rawVotes: String): Either<RatingValidationError, Rating> = catch {
             val rating = rawRating.toDouble()
-            val votes = rawVotes.replace(",", "").toLong()
+            val votes = rawVotes.replace(",", "").toInt()
             Rating(rating, votes)
         }.mapLeft { ex: Throwable ->
             log.error(ex) { ex.message }
